rename: `icon-trash` -> `delete` 파일명 변경
