+++ src/components/Input/QuizInputField.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { ReactComponent as CheckIcon } from '../../assets/icons/complete.svg';
 import { ReactComponent as EditIcon } from '../../assets/icons/edit_gray.svg';
-import { ReactComponent as TrashIcon } from '../../assets/icons/icon-trash.svg';
+import { ReactComponent as DeleteIcon } from '../../assets/icons/delete.svg';
 import { UserQuizInputType } from '../../types';
 
 const PLACEHOLDER = {
@@ -57,7 +57,7 @@ function QuizInputField({
           />
         )}
         {type === 'option' && handleDelete && (
-          <TrashIcon
+          <DeleteIcon
             className="icon"
             onClick={() => {
               handleDelete(index);
