+++ src/pages/Management/History/HistoryList/HistoryItem/index.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { styled } from 'styled-components';
 import FileApi from '../../../../../api/FileApi';
 import { ReactComponent as EditIcon } from '../../../../../assets/icons/edit_gray.svg';
-import { ReactComponent as DeleteIcon } from '../../../../../assets/icons/icon-trash.svg';
+import { ReactComponent as DeleteIcon } from '../../../../../assets/icons/delete.svg';
 import PDFButton from '../../../../../components/Button/PDFButton';
 import DeleteModal from '../../../../../components/Modal/DeleteModal';
 import Typography from '../../../../../components/Typography';
